Split module related classes out into their own files.
