+++ Blish HUD/Blish HUD.csproj
@@ -227,6 +227,13 @@
     <Compile Include="GameServices\Input\MouseEvent.cs" />
     <Compile Include="GameServices\Input\MouseEventArgs.cs" />
     <Compile Include="GameServices\Input\MouseEventType.cs" />
+    <Compile Include="GameServices\Modules\Manifest.cs" />
+    <Compile Include="GameServices\Modules\ModuleApiPermissions.cs" />
+    <Compile Include="GameServices\Modules\ModuleContributor.cs" />
+    <Compile Include="GameServices\Modules\ModuleDependency.cs" />
+    <Compile Include="GameServices\Modules\ModuleRunState.cs" />
+    <Compile Include="GameServices\Modules\ModuleRunStateChangedEventArgs.cs" />
+    <Compile Include="GameServices\Modules\SupportedModuleManifestVersion.cs" />
     <Compile Include="GameServices\PersistentStore\Store.cs" />
     <Compile Include="GameServices\PersistentStore\StoreValue.cs" />
     <Compile Include="GameServices\PersistentStore\StoreValue[T].cs" />
