+++ Blish HUD/GameServices/Modules/ModuleRunStateChangedEventArgs.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Blish_HUD.Modules {
+
+    public class ModuleRunStateChangedEventArgs : EventArgs {
+
+        public ModuleRunState RunState { get; }
+
+        public ModuleRunStateChangedEventArgs(ModuleRunState runState) {
+            this.RunState = runState;
+        }
+
+    }
+
+}
