+++ Blish HUD/GameServices/Modules/SupportedModuleManifestVersion.cs
@@ -1,0 +1,7 @@
+﻿namespace Blish_HUD.Modules {
+
+    public enum SupportedModuleManifestVersion : int {
+        V1 = 1,
+    }
+
+}
